# swagger-execise-lm.yaml
swagger: '2.0'
info:
  description: >-
    This is a sample server.
  version: 1.0.0
  title: Swagger Tax Calculator
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: execize@test.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: tax.io
basePath: /v1
tags:
  - name: tax
    description: endopoints about good taxes
  - name: category
    description: endpoints about good categories
schemes:
  - http
paths:
  /good/tax:
    post:
      tags:
        - tax
      summary: return given goods taxes
      description: ''
      operationId: getTaxes
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: good list for taxes computation 
          required: true
          schema:
            type: array
            items:
              $ref: '#/definitions/Good'
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/Order'
        '400':
          description: Invalid input
        '503':
          description: Fatal Error
  /good/category:
    get:
      tags:
        - category
      summary: Returns good categories
      description: Returns defined good categories 
      operationId: getCategories
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Category'
        '503':
          description: Fatal Error
definitions:
  Good:
    type: object
    properties:
      categoryId:
        type: integer
        format: int32
        description: category type id (as in good/category endpoint)
      quantity:
        type: integer
        format: int32
        description: good quantity
      price:
        type: number
        format: double
        description: single item price
      imported:
        type: boolean
        default: false
        description: imported flag
  Category:
    type: object
    properties:
      id:
        type: integer
        format: int32
        description: type category identifier
      description:
        type: string
        description: category description (ie food, general good)
      taxRate:
        type: number
        format: double
        description: tax rate applied to the category
      importTaxRate:
        type: number
        format: double
        description: import tax rate applied to the category
  Order:
    type: object
    properties:
      goods:
        type: array
        items:
          $ref: '#/definitions/Good'
        description: goods in order
      taxes:
        type: number
        format: double
        description: total taxes
      total:
        type: number
        format: double
        description: total amount